pyproject.toml
[project]
name = "validate-mypy-deps"
version = "0.1.0"
description = "Validate mypy pre-commit hook additional dependencies."
readme = "README.md"
authors = [{ name = "gleb.kharitonov", email = "haritonovgleb123@gmail.com" }]
requires-python = ">=3.12"
dependencies = ["ruamel-yaml>=0.18.16"]

[project.scripts]
validate-mypy-deps = "validate_mypy_deps.cli:main"

[build-system]
requires = ["uv_build>=0.8.19,<0.9.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["T201", "D"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"

[tool.pyright]
typeCheckingMode = "standard"

[dependency-groups]
lint = [
    "pre-commit>=4.3.0",
]
